Refactore utils.py module
IMPORTANT: The boot.py module must be updated to import from homie.network
instead of importing from homie.utils.

https://breadcrumbscollector.tech/stop-naming-your-python-modules-utils/
